Create default language if not exists

+++ src/Manager.php
@@ -64,10 +64,10 @@ class Manager
             return $this->translations;
         }
 
+        if(!$this->translations) $this->addLanguage(config('langmanGUI.base_language'));
+
         $this->getJsonTranslations();
         $this->getArrayTranslations();
-
-        if(!$this->translations) $this->addLanguage(config('langmanGUI.base_language'));
 
         return $this->translations;
     }
@@ -135,11 +135,15 @@ class Manager
     {
         $this->backup();
 
-        file_put_contents($this->languageFilesPath . DIRECTORY_SEPARATOR . "$language.json",
-            json_encode((object)[], JSON_UNESCAPED_UNICODE | JSON_PRETTY_PRINT)
-        );
+        if(!file_exists($this->languageFilesPath . DIRECTORY_SEPARATOR . "$language.json")) {
+            file_put_contents($this->languageFilesPath . DIRECTORY_SEPARATOR . "$language.json",
+                json_encode((object)['Hello World' => 'Hello World'], JSON_UNESCAPED_UNICODE | JSON_PRETTY_PRINT)
+            );
+        }
 
-        $this->disk->makeDirectory($this->languageFilesPath . DIRECTORY_SEPARATOR . "$language");
+        if(!file_exists($this->languageFilesPath . DIRECTORY_SEPARATOR . "$language")) {
+            $this->disk->makeDirectory($this->languageFilesPath . DIRECTORY_SEPARATOR . "$language");
+        }   
     }
 
     /**
